feat(image-url): fix console errors + header naming
