+++ package.json
@@ -22,7 +22,7 @@
     "@angular/platform-browser-dynamic": "^17.0.2",
     "@angular/router": "^17.0.2",
     "@aotw/components": "0.0.2",
-    "@aotw/core": "0.0.1",
+    "@aotw/core": "^0.0.2",
     "@aotw/ng-components": "^0.0.3-next.4",
     "@ngrx/effects": "^17.0.0",
     "@ngrx/router-store": "^17.0.0",
