+++ src/app/components/discover-header/discover-header.component.html
@@ -1,16 +1,15 @@
-<div class="info">
-  <aotw-ng-breadcrumb [items]="breadcrumb" (selectedItem)="goToPage($event)" />
-  <h2>{{ rangedTopic.type || '\u2014' }}</h2>
-  <div class="topic-name">
-    <button title="The Flag Archive Logo">
-      <aotw-ng-icon name="logo" />
-    </button>
-    <h3>{{ rangedTopic.name }}</h3>
-  </div>
-</div>
-<app-image
-  [src]="rangedTopic.imageUrl || (rangedTopic.id | image: 'assets/mock/images/' + archiveId: rangedTopic.rangeSuffix)"
-  [alt]="rangedTopic.name"
-  [alignToRight]="true"
-  [placeholder]="!rangedTopic.imageUrl && !rangedTopic.image"
-/>
+<h2>{{ mainTopicType }}</h2>
+<ul>
+  @for (topic of topics; track $index) {
+    <li>
+      <a
+        tabindex="0"
+        [ngClass]="{ active: topic.id === activeTopicId }"
+        (click)="setActiveTopic(topic.id)"
+        (keydown)="setActiveTopic(topic.id)"
+      >
+        {{ topic.name }}
+      </a>
+    </li>
+  }
+</ul>
