+++ src/app/components/discover-header/discover-header.component.scss
@@ -2,36 +2,39 @@
   display: grid;
   font-weight: var(--aotw-font-weight-regular);
   gap: var(--aotw-space-xs);
-  grid-template-columns: 1fr 7.5rem;
+  height: 7rem;
+  z-index: 1;
 }
 
-.info {
-  display: flex;
-  flex-direction: column;
-  gap: var(--aotw-space-s);
+h2 {
+  color: var(--aotw-color-white);
+  font-size: var(--aotw-font-size-s);
+  line-height: var(--aotw-font-line-height-s);
+  text-transform: uppercase;
+}
 
-  h2 {
-    color: var(--aotw-color-white);
-    font-size: var(--aotw-font-size-s);
-    line-height: var(--aotw-font-line-height-s);
-    text-transform: uppercase;
-  }
+ul {
+  display: grid;
+  font-size: var(--aotw-font-size-m);
+  grid-template-columns: repeat(2, 1fr);
+  line-height: var(--aotw-font-line-height-m);
+  list-style-type: none;
+  margin: 0;
+  padding: 0;
 
-  .topic-name {
-    display: flex;
-    gap: var(--aotw-space-s);
+  li {
+    a {
+      color: var(--aotw-color-white);
+      cursor: pointer;
+      text-decoration: none;
 
-    button {
-      aspect-ratio: 1 / 1;
-      border-radius: 50%;
-      height: var(--aotw-size-s);
-      justify-content: center;
-      padding: 0;
-    }
+      &:hover {
+        color: var(--aotw-color-primary-500);
+      }
 
-    h3 {
-      font-size: var(--aotw-font-size-m);
-      line-height: var(--aotw-font-line-height-s);
+      &.active {
+        color: var(--aotw-color-primary-500);
+      }
     }
   }
 }
