+++ src/app/components/discover-header/discover-header.component.ts
@@ -1,81 +1,44 @@
 import {
   ChangeDetectionStrategy,
-  ChangeDetectorRef,
   Component,
+  EventEmitter,
   Input,
   OnInit,
-  inject
+  Output
 } from '@angular/core';
-import { Router, RouterModule } from '@angular/router';
-import {
-  AotwBreadcrumbComponent,
-  AotwIconComponent,
-  BreadcrumbItem
-} from '@aotw/ng-components';
-import { Store } from '@ngrx/store';
-import { Subject, takeUntil } from 'rxjs';
 
-import { TopicWithRange } from '../../models';
-import { ImagePipe } from '../../pipes';
-import { TopicService } from '../../services';
+import { Topic } from '../../models';
 import { SharedModule } from '../../shared';
-import { selectSelectedYear } from '../../state/selectors';
-import { ImageComponent } from '../image';
 
 @Component({
   selector: 'app-discover-header',
   standalone: true,
-  imports: [
-    SharedModule,
-    RouterModule,
-    AotwBreadcrumbComponent,
-    AotwIconComponent,
-    ImageComponent,
-    ImagePipe
-  ],
+  imports: [ SharedModule ],
   changeDetection: ChangeDetectionStrategy.OnPush,
   templateUrl: './discover-header.component.html',
-  styleUrls: ['./discover-header.component.scss'],
+  styleUrl: './discover-header.component.scss'
 })
 export class DiscoverHeaderComponent implements OnInit {
   @Input()
-  public breadcrumb: BreadcrumbItem[] = [];
+  public activeTopicId?: string;
 
-  private _topic!: TopicWithRange;
-  public get topic(): TopicWithRange {
-    return this._topic;
-  }
   @Input()
-  public set topic(topic: TopicWithRange) {
-    this._topic = topic;
-    this.rangedTopic = this.topicService.setImageRange(topic);
-  }
+  public topics?: Topic[];
 
-  public archiveId!: string;
+  @Output()
+  public activeTopic = new EventEmitter<string>();
 
-  public rangedTopic = this.topic;
-
-  private cdr = inject(ChangeDetectorRef);
-  private router = inject(Router);
-  private topicService = inject(TopicService);
-  private store = inject(Store);
-  
-  private unsubscribe$ = new Subject<void>();
-  private selectSelectedYear$ = this.store.select(selectSelectedYear);
+  public mainTopicType?: string;
 
   public ngOnInit(): void {
-    this.archiveId = '23flag01';
+    if (!this.topics) {
+      return;
+    }
 
-    this.selectSelectedYear$.pipe(
-      takeUntil(this.unsubscribe$)
-    ).subscribe(selectedYear => {
-      this.rangedTopic = this.topicService.setImageRange(this.topic, selectedYear);
-      this.cdr.detectChanges();
-    });
+    this.mainTopicType = this.topics[0].type;
   }
 
-  public goToPage(item: BreadcrumbItem): void {
-    const route = item.link?.split('/');
-    this.router.navigate(route || []);
+  public setActiveTopic(id: string): void {
+    this.activeTopic.emit(id);
   }
 }
