+++ src/app/components/image/image.component.scss
@@ -3,9 +3,11 @@
   display: flex;
 
   img {
+    height: 100%;
     margin: auto 0;
     max-height: 100%;
     max-width: 100%;
+    width: 100%;
 
     &.right {
       margin: auto 0 auto auto;
