+++ src/app/components/topic-header/topic-header.component.scss
@@ -2,39 +2,36 @@
   display: grid;
   font-weight: var(--aotw-font-weight-regular);
   gap: var(--aotw-space-xs);
-  height: 7rem;
-  z-index: 1;
+  grid-template-columns: 1fr 7.5rem;
 }
 
-h2 {
-  color: var(--aotw-color-white);
-  font-size: var(--aotw-font-size-s);
-  line-height: var(--aotw-font-line-height-s);
-  text-transform: uppercase;
-}
+.info {
+  display: flex;
+  flex-direction: column;
+  gap: var(--aotw-space-s);
 
-ul {
-  display: grid;
-  font-size: var(--aotw-font-size-m);
-  grid-template-columns: repeat(2, 1fr);
-  line-height: var(--aotw-font-line-height-m);
-  list-style-type: none;
-  margin: 0;
-  padding: 0;
+  h2 {
+    color: var(--aotw-color-white);
+    font-size: var(--aotw-font-size-s);
+    line-height: var(--aotw-font-line-height-s);
+    text-transform: uppercase;
+  }
 
-  li {
-    a {
-      color: var(--aotw-color-white);
-      cursor: pointer;
-      text-decoration: none;
+  .topic-name {
+    display: flex;
+    gap: var(--aotw-space-s);
 
-      &:hover {
-        color: var(--aotw-color-primary-500);
-      }
+    button {
+      aspect-ratio: 1 / 1;
+      border-radius: 50%;
+      height: var(--aotw-size-s);
+      justify-content: center;
+      padding: 0;
+    }
 
-      &.active {
-        color: var(--aotw-color-primary-500);
-      }
+    h3 {
+      font-size: var(--aotw-font-size-m);
+      line-height: var(--aotw-font-line-height-s);
     }
   }
 }
