+++ src/main.ts
@@ -1,3 +1,4 @@
+import { IMAGE_CONFIG } from '@angular/common';
 import { HttpClient, provideHttpClient } from '@angular/common/http';
 import { importProvidersFrom } from '@angular/core';
 import { bootstrapApplication } from '@angular/platform-browser';
@@ -39,6 +40,15 @@ bootstrapApplication(AppComponent, {
       maxAge: 25,
       trace: false,
       traceLimit: 75
-    })
+    }),
+    // ? - Needed to surpress warnings about intrinsic sizing for external SVG sources
+    // TODO - Find solution to decrease intrinsic size of external SVG sources
+    {
+      provide: IMAGE_CONFIG,
+      useValue: {
+        disableImageSizeWarning: true, 
+        disableImageLazyLoadWarning: true
+      }
+    },
   ]
 }).catch(err => console.error(err));
